django_backend/backend/requirements.txt
absl-py==0.11.0
asgiref==3.2.10
astor==0.8.1
cached-property==1.5.2
cycler==0.10.0
decorator==4.4.2
Django==3.1.2
djangorestframework==3.12.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.9.0
imageio==2.9.0
importlib-metadata==2.0.0
jsonify==0.5
Keras==2.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.0.3
mock==4.0.2
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.20.3
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Theano==1.0.5
tifffile==2020.11.18
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
